+++ crates/rpc-types-any/Cargo.toml
@@ -24,12 +24,9 @@ workspace = true
 
 [dependencies]
 alloy-consensus-any = { workspace = true, features = ["serde"] }
-alloy-rpc-types-eth.workspace = true
-alloy-serde.workspace = true
-
-[dev-dependencies]
+alloy-network-primitives.workspace = true
 alloy-primitives.workspace = true
 alloy-rpc-types-eth = { workspace = true, features = ["serde"] }
-
+alloy-serde.workspace = true
 serde.workspace = true
 serde_json.workspace = true
